refactor: Rename models to model
